Clean up the public API file and improve a few other things
