+++ lib/WebpackConfig.js
@@ -12,14 +12,19 @@
 const path = require('path');
 const fs = require('fs');
 const crypto = require('crypto');
+const RuntimeConfig = require('./config/RuntimeConfig'); //eslint-disable-line no-unused-vars
 const logger = require('./logger');
 
 /**
- * @param {RuntimeConfig} runtimeConfig
+ * @param {RuntimeConfig|null} runtimeConfig
  * @return {void}
  */
 function validateRuntimeConfig(runtimeConfig) {
     // if you're using the encore executable, these things should never happen
+    if (null === runtimeConfig) {
+        throw new Error('RuntimeConfig must be initialized');
+    }
+
     if (null === runtimeConfig.context) {
         throw new Error('RuntimeConfig.context must be set.');
     }
@@ -29,9 +34,43 @@ function validateRuntimeConfig(runtimeConfig) {
     }
 }
 
+/**
+ * @param {RuntimeConfig} runtimeConfig
+ * @return {void}
+ */
+function checkPackageJson(runtimeConfig) {
+    // Display a warning if webpack is listed as a [dev-]dependency
+    try {
+        const packageInfo = JSON.parse(
+            fs.readFileSync(path.resolve(runtimeConfig.context, 'package.json'), { encoding: 'utf-8' })
+        );
+
+        if (packageInfo) {
+            const dependencies = new Set([
+                ...(packageInfo.dependencies ? Object.keys(packageInfo.dependencies) : []),
+                ...(packageInfo.devDependencies ? Object.keys(packageInfo.devDependencies) : []),
+            ]);
+
+            if (dependencies.has('webpack')) {
+                logger.warning('Webpack is already provided by Webpack Encore, also adding it to your package.json file may cause issues.');
+            }
+        }
+    } catch (e) {
+        logger.warning('Could not read package.json file.');
+    }
+}
+
 class WebpackConfig {
-    constructor(runtimeConfig) {
+    constructor(runtimeConfig, enablePackageJsonCheck = false) {
         validateRuntimeConfig(runtimeConfig);
+
+        if (runtimeConfig.verbose) {
+            logger.verbose();
+        }
+
+        if (enablePackageJsonCheck) {
+            checkPackageJson(runtimeConfig);
+        }
 
         this.runtimeConfig = runtimeConfig;
         this.entries = new Map();
@@ -50,6 +89,8 @@ class WebpackConfig {
         this.aliases = {};
         this.externals = [];
         this.integrityAlgorithms = [];
+        this.shouldUseSingleRuntimeChunk = null;
+        this.shouldSplitEntryChunks = false;
 
         // Features/Loaders flags
         this.useVersioning = false;
@@ -96,8 +137,6 @@ class WebpackConfig {
         this.stylusLoaderOptionsCallback = () => {};
         this.babelConfigurationCallback = () => {};
         this.cssLoaderConfigurationCallback = () => {};
-        this.shouldUseSingleRuntimeChunk = null;
-        this.shouldSplitEntryChunks = false;
         this.splitChunksConfigurationCallback = () => {};
         this.watchOptionsConfigurationCallback = () => {};
         this.vueLoaderOptionsCallback = () => {};
