+++ lib/config-generator.js
@@ -9,6 +9,7 @@
 
 'use strict';
 
+const WebpackConfig = require('./WebpackConfig'); //eslint-disable-line no-unused-vars
 const cssExtractLoaderUtil = require('./loaders/css-extract');
 const pathUtil = require('./config/path-util');
 const loaderFeatures = require('./features');
