+++ lib/config/RuntimeConfig.js
@@ -16,11 +16,13 @@ class RuntimeConfig {
         this.isValidCommand = false;
         this.environment = process.env.NODE_ENV ? process.env.NODE_ENV : 'dev';
 
-        this.useDevServer = null;
+        this.useDevServer = false;
         this.devServerUrl = null;
         this.devServerHttps = null;
         this.devServerKeepPublicPath = false;
-        this.useHotModuleReplacement = null;
+        this.useHotModuleReplacement = false;
+        this.outputJson = false;
+        this.profile = false;
 
         this.babelRcFileExists = null;
