+++ lib/config/parse-runtime.js
@@ -22,10 +22,6 @@ const babel = require('@babel/core');
 module.exports = function(argv, cwd) {
     const runtimeConfig = new RuntimeConfig();
     runtimeConfig.command = argv._[0];
-    runtimeConfig.useDevServer = false;
-    runtimeConfig.useHotModuleReplacement = false;
-    runtimeConfig.outputJson = false;
-    runtimeConfig.profile = false;
 
     switch (runtimeConfig.command) {
         case 'dev':
