+++ lib/config/path-util.js
@@ -10,6 +10,7 @@
 'use strict';
 
 const path = require('path');
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 
 module.exports = {
     /**
