+++ lib/config/validator.js
@@ -11,6 +11,7 @@
 
 const pathUtil = require('./path-util');
 const logger = require('./../logger');
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 
 class Validator {
     /**
