+++ lib/friendly-errors/asset-output-display-plugin.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const chalk = require('chalk');
+const chalk = require('chalk').default;
 
 function AssetOutputDisplayPlugin(outputPath, friendlyErrorsPlugin) {
     this.outputPath = outputPath;
