+++ lib/friendly-errors/formatters/missing-css-file.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const chalk = require('chalk');
+const chalk = require('chalk').default;
 
 function formatErrors(errors) {
     if (errors.length === 0) {
