+++ lib/friendly-errors/formatters/missing-loader.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const chalk = require('chalk');
+const chalk = require('chalk').default;
 const loaderFeatures = require('../../features');
 
 function formatErrors(errors) {
