+++ lib/loaders/eslint.js
@@ -9,14 +9,15 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const loaderFeatures = require('../features');
 const applyOptionsCallback = require('../utils/apply-options-callback');
 
-/**
- * @param {WebpackConfig} webpackConfig
- * @return {Object} of options to use for eslint-loader options.
- */
 module.exports = {
+    /**
+     * @param {WebpackConfig} webpackConfig
+     * @return {Object} of options to use for eslint-loader options.
+     */
     getOptions(webpackConfig) {
         loaderFeatures.ensurePackagesExistAndAreCorrectVersion('eslint');
