+++ lib/loaders/handlebars.js
@@ -9,14 +9,15 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const loaderFeatures = require('../features');
 const applyOptionsCallback = require('../utils/apply-options-callback');
 
-/**
- * @param {WebpackConfig} webpackConfig
- * @return {Array} of loaders to use for Handlebars
- */
 module.exports = {
+    /**
+     * @param {WebpackConfig} webpackConfig
+     * @return {Array} of loaders to use for Handlebars
+     */
     getLoaders(webpackConfig) {
         loaderFeatures.ensurePackagesExistAndAreCorrectVersion('handlebars');
