+++ lib/loaders/stylus.js
@@ -9,16 +9,17 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const loaderFeatures = require('../features');
 const cssLoader = require('./css');
 const applyOptionsCallback = require('../utils/apply-options-callback');
 
-/**
- * @param {WebpackConfig} webpackConfig
- * @param {bool} useCssModules
- * @return {Array} of loaders to use for Stylus files
- */
 module.exports = {
+    /**
+     * @param {WebpackConfig} webpackConfig
+     * @param {boolean} useCssModules
+     * @return {Array} of loaders to use for Stylus files
+     */
     getLoaders(webpackConfig, useCssModules = false) {
         loaderFeatures.ensurePackagesExistAndAreCorrectVersion('stylus');
