+++ lib/loaders/typescript.js
@@ -9,15 +9,16 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const loaderFeatures = require('../features');
 const babelLoader = require('./babel');
 const applyOptionsCallback = require('../utils/apply-options-callback');
 
-/**
- * @param {WebpackConfig} webpackConfig
- * @return {Array} of loaders to use for TypeScript
- */
 module.exports = {
+    /**
+     * @param {WebpackConfig} webpackConfig
+     * @return {Array} of loaders to use for TypeScript
+     */
     getLoaders(webpackConfig) {
         loaderFeatures.ensurePackagesExistAndAreCorrectVersion('typescript');
