+++ lib/logger.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const chalk = require('chalk');
+const chalk = require('chalk').default;
 
 const messagesKeys = [
     'debug',
