+++ lib/package-helper.js
@@ -9,7 +9,7 @@
 
 'use strict';
 
-const chalk = require('chalk');
+const chalk = require('chalk').default;
 const fs = require('fs');
 const logger = require('./logger');
 const semver = require('semver');
@@ -25,7 +25,7 @@ ${missingPackagesRecommendation.message}
     }
 
     // check for invalid versions & warn
-    const invalidVersionRecommendations = getInvalidPackageVersionRecommendations(packagesConfig, requestedFeature);
+    const invalidVersionRecommendations = getInvalidPackageVersionRecommendations(packagesConfig);
     for (let message of invalidVersionRecommendations) {
         logger.warning(message);
     }
