+++ lib/plugins/clean.js
@@ -9,6 +9,7 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 const PluginPriorities = require('./plugin-priorities');
 const applyOptionsCallback = require('../utils/apply-options-callback');
