+++ lib/plugins/define.js
@@ -10,6 +10,7 @@
 'use strict';
 
 const webpack = require('webpack');
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const PluginPriorities = require('./plugin-priorities');
 const applyOptionsCallback = require('../utils/apply-options-callback');
