+++ lib/plugins/delete-unused-entries.js
@@ -9,6 +9,7 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const DeleteUnusedEntriesJSPlugin = require('../webpack/delete-unused-entries-js-plugin');
 const PluginPriorities = require('./plugin-priorities');
 const copyEntryTmpName = require('../utils/copyEntryTmpName');
