+++ lib/plugins/friendly-errors.js
@@ -9,6 +9,7 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
 const missingCssFileTransformer = require('../friendly-errors/transformers/missing-css-file');
 const missingCssFileFormatter = require('../friendly-errors/formatters/missing-css-file');
