+++ lib/plugins/mini-css-extract.js
@@ -9,6 +9,7 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const PluginPriorities = require('./plugin-priorities');
