+++ lib/plugins/optimize-css-assets.js
@@ -9,6 +9,7 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const applyOptionsCallback = require('../utils/apply-options-callback');
