+++ lib/plugins/shared-entry-concat.js
@@ -9,6 +9,7 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const SharedEntryConcatPlugin = require('../webpack/shared-entry-concat-plugin');
 const PluginPriorities = require('./plugin-priorities');
