+++ lib/plugins/terser.js
@@ -9,6 +9,7 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const TerserPlugin = require('terser-webpack-plugin');
 const applyOptionsCallback = require('../utils/apply-options-callback');
