+++ lib/plugins/variable-provider.js
@@ -9,13 +9,14 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const webpack = require('webpack');
 const PluginPriorities = require('./plugin-priorities');
 
 /**
  * @param {Array} plugins
  * @param {WebpackConfig} webpackConfig
- * @return {Array} of plugins to add to webpack
+ * @return {void}
  */
 module.exports = function(plugins, webpackConfig) {
     if (Object.keys(webpackConfig.providedVariables).length > 0) {
