+++ lib/plugins/versioning.js
@@ -11,6 +11,7 @@
 
 const webpack = require('webpack');
 const WebpackChunkHash = require('webpack-chunk-hash');
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
 const PluginPriorities = require('./plugin-priorities');
 
 /**
