+++ lib/utils/manifest-key-prefix-helper.js
@@ -9,6 +9,8 @@
 
 'use strict';
 
+const WebpackConfig = require('../WebpackConfig'); //eslint-disable-line no-unused-vars
+
 /**
  * Helper for determining the manifest.json key prefix.
  *
